rich text migration clean up
