+++ config.d.ts
@@ -4,7 +4,6 @@ declare module 'react-native-config' {
     REACT_APP_ENDPOINT: string;
     REACT_APP_POSTS_DATABASE: string;
     REACT_APP_POSTS_COLLECTION: string;
-    REACT_APP_POSTS_DATA_COLLECTION: string;
     REACT_APP_LOGIN_LOCATION_COLLECTION: string;
     REACT_APP_POSTS_BUCKET: string;
     REACT_APP_GEO_LOGIN_TOKEN: string;
