+++ global.d.ts
@@ -5,7 +5,6 @@ declare namespace NodeJS {
     REACT_APP_POSTS_DATABASE: string;
     REACT_APP_POSTS_COLLECTION: string;
     REACT_APP_POSTS_BUCKET: string;
-    REACT_APP_POSTS_DATA_COLLECTION: string;
     REACT_APP_LOGIN_LOCATION_COLLECTION: string;
     REACT_APP_GEO_LOGIN_TOKEN: string;
     REACT_APP_POSTS_METRICS_COLLECTION: string;
