+++ src/constants/Screens.tsx
@@ -1,8 +1,7 @@
 export const selfInfoScreen: string = 'SelfInfoScreen';
 export const loginScreen: string = 'LoginScreen';
 export const dashboardScreen: string = 'DashboardScreen';
-export const addPostScreen: string = 'AddPostScreen';
-export const postContentScreenV2: string = 'PostContentScreenV2';
+export const postContentScreen: string = 'PostContentScreen';
 export const privacyScreen: string = 'PrivacyScreen';
 export const contactScreen: string = 'ContactScreen';
 export const verificationEmailScreen: string = 'VerificationEmailScreen';
