+++ src/screens/Post/PostContentScreen.tsx
@@ -1,17 +1,14 @@
 import React, {useEffect, useState} from 'react';
 import {
   ActivityIndicator,
-  FlatList,
   StyleSheet,
   TouchableOpacity,
   View,
 } from 'react-native';
-import {PostContent, Theme} from '../../constants/Types';
-import NewPostComponent from '../../components/post/NewPostComponent';
-import {BUTTON_TYPES, TEXT_POST_TYPE} from '../../constants/Constants';
+import {Theme} from '../../constants/Types';
+import {BUTTON_TYPES} from '../../constants/Constants';
 import postService from '../../appwrite/posts';
 import {useModal} from '../../context/modal/useModal';
-import PostComponent from '../../components/post/PostComponent';
 import CustomText from '../../components/common/CustomText';
 import Button from '../../components/common/Button';
 import Wrapper from '../../components/common/Wrapper';
@@ -21,7 +18,6 @@ import PostStatusButton from '../../components/post/PostStatusButton';
 import Status from '../../components/post/enum/PostStatusEnum';
 import VideoUrlComponent from '../../components/post/VideoUrlComponent';
 import TLDRComponent from '../../components/post/TLDRComponent';
-import strings from '../../constants/strings.json';
 import Icon from '../../assets/Icon';
 import {ParamListBase, useNavigation} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
@@ -30,16 +26,21 @@ import GithubLink from '../../components/post/GithubLink';
 import {formatDate} from '../../helpers/functions';
 import {PostMetrics} from '../../appwrite/types/post_metrics';
 import postMetricsService from '../../appwrite/postMetrics';
+import RichTextEditor from '../../components/common/RichTextEditor';
+import {Post} from '../../appwrite/types/posts';
+import HtmlRenderer from '../../components/post/HtmlRenderer';
 
 function PostContentScreen({route}: any): JSX.Element {
   const navigation = useNavigation<NativeStackNavigationProp<ParamListBase>>();
-  const [post, setPost] = useState(route.params);
+  const [post, setPost] = useState<Post>(route.params);
   const [postMetrics, setPostMetrics] = useState<PostMetrics>();
-  const [newPostData, setNewPostData] = useState<PostContent | null>(null);
   const {isAdmin} = useUser();
 
   const [loading, setLoading] = useState(false);
-  const [updatingPost, setUpdatingPost] = useState(false);
+  const [isEditing, setIsEditing] = useState(
+    !post.content || post.content.length === 0 ? true : false,
+  );
+  const [showMenu, setShowMenu] = useState(false);
 
   const {theme} = useTheme();
 
@@ -50,50 +51,27 @@ function PostContentScreen({route}: any): JSX.Element {
         setPostMetrics(response[0]);
       }
     });
+    getPost();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [route]);
-
-  const onChange = (value: PostContent) => {
-    setNewPostData(value);
-  };
 
   const {openModal} = useModal();
 
-  const onAdd = () => {
-    setNewPostData({
-      title: '',
-      subtitle: '',
-      content_type: TEXT_POST_TYPE,
-      content: '',
-      postID: post.$id,
-    });
-  };
-
-  const onSave = async () => {
-    if (newPostData) {
-      setUpdatingPost(true);
-      await postService
-        .createPostContent(newPostData, post)
-        .then(() => {
-          setUpdatingPost(false);
-          getPost();
-          setNewPostData(null);
-        })
-        .catch(err => {
-          setUpdatingPost(false);
-          if (err instanceof Error) {
-            openModal({title: err.message});
-          } else {
-            openModal({title: 'Unknown error occurred'});
-          }
-        });
-    }
+  const onChangePostContent = (value: string) => {
+    setPost(prev => ({
+      ...prev,
+      content: value,
+    }));
   };
 
   const getPost = async () => {
+    if (!post.$id) {
+      return;
+    }
     await postService
-      .getPost(post?.$id)
+      .getPost(post.$id)
       .then(response => {
-        setPost(response);
+        setPost(response as unknown as Post);
       })
       .catch(err => {
         if (err instanceof Error) {
@@ -104,12 +82,25 @@ function PostContentScreen({route}: any): JSX.Element {
       });
   };
 
+  const onPressSavePost = async () => {
+    setLoading(true);
+    setIsEditing(false);
+    await postService
+      .updatePost(post?.$id ?? '', post)
+      .then(() => {
+        console.log('Post saved successfully');
+        // setPost(response as unknown as Post);
+      })
+      .catch(err => openModal({title: err?.message}));
+    setLoading(false);
+  };
+
   const onPostStatusChange = async (status: Status) => {
     setLoading(true);
     await postService
       .updatePost(post?.$id ?? '', {...post, status: status})
       .then(response => {
-        setPost(response);
+        setPost(response as unknown as Post);
       })
       .catch(err => openModal({title: err?.message}));
     setLoading(false);
@@ -120,7 +111,7 @@ function PostContentScreen({route}: any): JSX.Element {
     await postService
       .updatePost(post?.$id ?? '', {...post, videoUrl: url})
       .then(response => {
-        setPost(response);
+        setPost(response as unknown as Post);
       })
       .catch(err => openModal({title: err?.message}));
     setLoading(false);
@@ -132,7 +123,7 @@ function PostContentScreen({route}: any): JSX.Element {
       .updatePost(post?.$id ?? '', {...post, tldr: value})
       .then(response => {
         console.log(response);
-        setPost(response);
+        setPost(response as unknown as Post);
       })
       .catch(err => openModal({title: err?.message}));
     setLoading(false);
@@ -144,7 +135,7 @@ function PostContentScreen({route}: any): JSX.Element {
       .updatePost(post?.$id ?? '', {...post, githubUrl: value})
       .then(response => {
         console.log(response);
-        setPost(response);
+        setPost(response as unknown as Post);
       })
       .catch(err => openModal({title: err?.message}));
     setLoading(false);
@@ -155,16 +146,85 @@ function PostContentScreen({route}: any): JSX.Element {
     openModal({title: 'Link Copied, You can share it now.'});
   };
 
+  const onPressEdit = () => {
+    setIsEditing(true);
+    setShowMenu(false);
+  };
+
   return (
-    <Wrapper style={styles(theme).container}>
+    <Wrapper
+      showsVerticalScrollIndicator={true}
+      contentAboveScrollable={
+        <View style={styles(theme).header}>
+          <TouchableOpacity onPress={() => navigation.goBack()}>
+            <Icon
+              icon={'arrow-left2'}
+              size={theme.sizes.large}
+              color={theme.colors.text_color}
+            />
+          </TouchableOpacity>
+          {isAdmin && (
+            <View style={styles(theme).headerRight}>
+              <View>
+                <Button
+                  buttonStyle={styles(theme).saveButtonStyle}
+                  title="Share"
+                  iconLeft={
+                    <Icon
+                      containerStyle={styles(theme).shareIconContainerStyle}
+                      icon={'share'}
+                      size={theme.sizes.medium}
+                      color={theme.colors.button_text}
+                    />
+                  }
+                  type={BUTTON_TYPES.outlined}
+                  onPress={onPressShare}
+                />
+              </View>
+              {isEditing && (
+                <View>
+                  <Button
+                    buttonStyle={styles(theme).saveButtonStyle}
+                    title="Save"
+                    type={BUTTON_TYPES.filled}
+                    onPress={onPressSavePost}
+                    loading={loading}
+                  />
+                </View>
+              )}
+              {post.status !== Status.published && (
+                <View>
+                  <Button
+                    buttonStyle={styles(theme).saveButtonStyle}
+                    title="Publish"
+                    type={BUTTON_TYPES.filled}
+                    onPress={() => onPostStatusChange(Status.published)}
+                    loading={loading}
+                  />
+                </View>
+              )}
+              <View>
+                <TouchableOpacity onPress={() => setShowMenu(!showMenu)}>
+                  <Icon
+                    icon={'menu'}
+                    size={theme.sizes.large}
+                    color={theme.colors.text_color}
+                  />
+                </TouchableOpacity>
+                {showMenu && (
+                  <View style={styles(theme).menuStyle}>
+                    <TouchableOpacity onPress={onPressEdit}>
+                      <CustomText title={'Edit'} type={'p1'} />
+                    </TouchableOpacity>
+                  </View>
+                )}
+              </View>
+            </View>
+          )}
+        </View>
+      }
+      style={styles(theme).container}>
       <View style={styles(theme).titleContainer}>
-        <TouchableOpacity onPress={() => navigation.goBack()}>
-          <Icon
-            icon={'arrow-left2'}
-            size={theme.sizes.extra_extra_large}
-            color={theme.colors.text_color}
-          />
-        </TouchableOpacity>
         <View style={styles(theme).titleWithTime}>
           <CustomText
             style={styles(theme).postTitle}
@@ -174,7 +234,10 @@ function PostContentScreen({route}: any): JSX.Element {
           <View style={styles(theme).timeAndViews}>
             <CustomText
               style={styles(theme).date}
-              title={' · ' + formatDate(new Date(post?.$createdAt))}
+              title={
+                ' · ' +
+                formatDate(new Date(post?.$createdAt as unknown as string))
+              }
               type={'p2'}
             />
             {postMetrics && (
@@ -204,61 +267,37 @@ function PostContentScreen({route}: any): JSX.Element {
           status={post.status ? post.status : Status.pending}
         />
       )}
-      <GithubLink
-        loading={loading}
-        onChange={newUrl => onGithubURLUpdate(newUrl)}
-        url={post.githubUrl}
-      />
-      <VideoUrlComponent
-        loading={loading}
-        url={post?.videoUrl}
-        onUrlChange={(url: string) => onPostVideoUrlChange(url)}
-      />
-      {post.contents.length === 0 && (
-        <CustomText title={strings.startByAdding} type={'p1'} />
-      )}
-      <View style={styles(theme).headerContainer}>
-        <FlatList
-          showsVerticalScrollIndicator={false}
-          data={post.contents}
-          renderItem={({item}) => <PostComponent id={item} />}
-          keyExtractor={(item, index) => index.toString()}
-        />
-      </View>
-      {newPostData && (
-        <NewPostComponent
-          loading={updatingPost}
-          newPost={newPostData}
-          onChange={onChange}
-          onSave={onSave}
+      {(post.githubUrl || isEditing) && (
+        <GithubLink
+          loading={loading}
+          onChange={newUrl => onGithubURLUpdate(newUrl)}
+          url={post.githubUrl as unknown as string}
         />
       )}
-      {!newPostData && isAdmin && (
-        <Button
-          buttonStyle={styles(theme).buttonStyle}
-          title="Add Content"
-          type={BUTTON_TYPES.filled}
-          onPress={onAdd}
+      {(post?.videoUrl || isEditing) && (
+        <VideoUrlComponent
+          loading={loading}
+          url={post?.videoUrl as unknown as string}
+          onUrlChange={(url: string) => onPostVideoUrlChange(url)}
         />
       )}
-      <TLDRComponent
-        loading={loading}
-        onChange={value => onPostTLDRUpdate(value)}
-        content={post?.tldr}
-      />
-      <Button
-        buttonStyle={styles(theme).buttonStyle}
-        title="Share"
-        type={BUTTON_TYPES.filled}
-        onPress={onPressShare}
-        iconRight={
-          <Icon
-            icon="share2"
-            size={theme.sizes.large}
-            color={theme.colors.button_text_filled}
+      {(post?.tldr || isEditing) && (
+        <TLDRComponent
+          loading={loading}
+          onChange={value => onPostTLDRUpdate(value)}
+          content={post?.tldr as unknown as string}
+        />
+      )}
+      <View style={styles(theme).contentContainer}>
+        {isEditing ? (
+          <RichTextEditor
+            onChangeText={onChangePostContent}
+            value={post.content}
           />
-        }
-      />
+        ) : (
+          <HtmlRenderer content={post.content} />
+        )}
+      </View>
     </Wrapper>
   );
 }
@@ -266,6 +305,7 @@ const styles = (theme: Theme) =>
   StyleSheet.create({
     container: {
       paddingBottom: theme.sizes.extra_extra_large * 2,
+      paddingTop: theme.sizes.large,
     },
     headerContainer: {
       marginBottom: theme.sizes.medium,
@@ -275,6 +315,8 @@ const styles = (theme: Theme) =>
     },
     buttonStyle: {
       alignSelf: 'center',
+      marginBottom: theme.sizes.medium,
+      marginTop: theme.sizes.extra_extra_large,
     },
     indicator: {
       marginTop: theme.sizes.medium,
@@ -291,6 +333,46 @@ const styles = (theme: Theme) =>
     },
     date: {
       marginRight: theme.sizes.small,
+    },
+    header: {
+      position: 'absolute',
+      top: 0,
+      left: 0,
+      flex: 1,
+      width: '100%',
+      flexDirection: 'row',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      paddingHorizontal: theme.sizes.large,
+      paddingVertical: theme.sizes.extra_small,
+      zIndex: 10,
+      backgroundColor: theme.colors.background_color,
+    },
+    saveButtonStyle: {
+      paddingVertical: theme.sizes.extra_small,
+      paddingHorizontal: theme.sizes.small,
+    },
+    headerRight: {
+      alignContent: 'flex-end',
+      flexDirection: 'row',
+      alignItems: 'center',
+      gap: theme.sizes.extra_small,
+    },
+    menuStyle: {
+      position: 'absolute',
+      bottom: -40,
+      left: -50,
+      backgroundColor: theme.colors.background_color,
+      padding: theme.sizes.small,
+      width: 80,
+      alignItems: 'flex-end',
+      borderRadius: theme.sizes.border_radius,
+    },
+    shareIconContainerStyle: {
+      padding: 0,
+    },
+    contentContainer: {
+      marginTop: theme.sizes.medium,
     },
   });
